+++ transpiler/main/code_gen/c_code_gen/c_code_gen.c
@@ -21,6 +21,7 @@
 #include "code_gen/structs/structs_code_gen.h"
 
 #include "analyzer/fn_analyzer.h"
+#include "analyzer/dead_analyzer.h"
 
 #include "tables/sst/sst.h"
 #include "tables/sst/sst_prefill.h"
@@ -83,9 +84,12 @@ static void fill_tables(struct AST* ast, struct Ctx* ctx){
 	sst_clear(ctx->tables->sst);
 	sst_prefill(ctx->tables, ctx->tables->sst);
 	
+	
 	for(int i = 0; i < ast->count_namespaces; i++){
 		
 		struct Namespace* ns = ast->namespaces[i];
+		
+		add_gen_struct_subrs_sst(ctx, ns);
 		
 		sst_fill(ctx->tables->sst, ns, ctx->flags->debug);
 		stst_fill(ctx->tables->stst, ns, ctx->flags->debug);
@@ -112,14 +116,10 @@ static void transpileAST(struct AST* ast, struct Ctx* ctx){
 	
 	ctx->flags->has_main_fn = sst_contains(ctx->tables->sst, "main");
 	
-	//TODO: re-enable
-	//analyze_functions(ctx->tables, ast);
 	
-	//TODO: instead of transpiling namespaces
-	//separately, the forward declarations of all
-	//namespaces should be transpiled
-	//before any functions/structures
-	//are defined
+	analyze_functions(ctx->tables, ast, ctx->flags->debug);
+	analyze_dead_code(ctx->tables, ast);
+	
 	
 	for(int i=0; i < ast->count_namespaces; i++)
 	{ ns_transpile_fwd(ast->namespaces[i], ctx); }
@@ -136,7 +136,9 @@ static void ns_transpile_struct_fwd_decls(struct Namespace* ns, struct Ctx* ctx)
 	//write struct fwd. declarations
 	for(int i=0;i < ns->count_structs; i++){
 		
-		char* name = ns->structs[i]
+		struct StructDecl* decl = ns->structs[i];
+		
+		char* name = decl
 					   ->type
 			           ->structType
 					   ->typeName;
@@ -155,7 +157,13 @@ static void ns_transpile_struct_decls(struct Namespace* ns, struct Ctx* ctx){
 static void ns_transpile_subr_fwd_decls(struct Namespace* ns, struct Ctx* ctx){
 
 	for(int i=0; i < ns->count_methods; i++){
-		transpileMethodSignature(ns->methods[i], ctx);
+		
+		struct Method* m = ns->methods[i];
+		
+		if(sst_get(ctx->tables->sst, m->name)->is_dead)
+			{continue; }
+		
+		transpileMethodSignature(m, ctx);
 		fprintf(ctx->file, ";\n");
 	}
 }
@@ -163,7 +171,13 @@ static void ns_transpile_subr_fwd_decls(struct Namespace* ns, struct Ctx* ctx){
 static void ns_transpile_subrs(struct Namespace* ns, struct Ctx* ctx){
 
 	for(int i=0; i < ns->count_methods; i++){
-		transpileMethod(ns->methods[i], ctx);
+		
+		struct Method* m = ns->methods[i];
+		
+		if(sst_get(ctx->tables->sst, m->name)->is_dead)
+			{continue; }
+		
+		transpileMethod(m, ctx);
 	}
 }