+++ transpiler/main/code_gen/structs/structs_code_gen.h
@@ -8,4 +8,6 @@ struct Ctx;
 void gen_struct_subrs(struct Namespace* ns, struct Ctx* ctx);
 void gen_struct_subr_signatures(struct Namespace* ns, struct Ctx* ctx);
 
+void add_gen_struct_subrs_sst(struct Ctx* ctx, struct Namespace* ns);
+
 #endif